python_sdk/requirements.txt
PySide6-Addons==6.4.0.1
PySide6-Essentials==6.4.0.1
PySide6==6.4.0.1
loguru==0.5.3
numpy==1.24.4
onnxruntime-gpu==1.13.1
onnxruntime==1.10.0
opencv-contrib-python==4.8.0.74
opencv-python==4.6.0.66
pupil-apriltags==1.0.4.post10
pyrealsense2==2.51.1.4348
pyserial==3.5
pywin32==300
pyzbar==0.1.9
scipy==1.10.1
simple-pid==2.0.0
